Refactor to internal module
