--- lib/provider/ui/windows/solitaire/foudation-atom.ts
@@ -1,10 +1,0 @@
-import type { Card } from "~/lib/solitaire-card-interface";
-
-import { atom } from "jotai";
-
-export const foundationAtom = atom<Record<number, Card[]>>({
-  0: [],
-  1: [],
-  2: [],
-  3: [],
-});
