+++ lib/provider/ui/windows/solitaire/index.tsx
@@ -4,29 +4,22 @@ import { useSetAtom } from "jotai";
 import { Button, Toolbar } from "react95";
 import styled from "styled-components";
 
-import { stock } from "~/lib/solitaire-global";
 import { Window } from "~/lib/window";
 
-import { foundationAtom } from "./foudation-atom";
+import { foundationAtom, tableauAtom, wasteAtom } from "./atom";
 import { Game, init } from "./game";
-import { tableauAtom } from "./tableau-atom";
-import { wasteAtom } from "./waste-atom";
+import { stock } from "./type";
 
 export function Solitaire(): ReactNode {
   const setWaste = useSetAtom(wasteAtom);
   const setFoundation = useSetAtom(foundationAtom);
   const setTableau = useSetAtom(tableauAtom);
 
-  const resetGame = (): void => {
+  const handleNewGame = (): void => {
     setWaste([]);
     setFoundation({ 0: [], 1: [], 2: [], 3: [] });
     stock.length = 0;
-  };
-
-  const handleNewGame = (): void => {
-    resetGame();
-    const tableau = init();
-    setTableau(tableau);
+    init(setTableau);
   };
 
   return (
