+++ lib/provider/ui/windows/solitaire/spritesheet.png
