--- lib/provider/ui/windows/solitaire/waste-atom.ts
@@ -1,5 +1,0 @@
-import type { Card } from "~/lib/solitaire-card-interface";
-
-import { atom } from "jotai";
-
-export const wasteAtom = atom<Card[]>([]);
