--- lib/solitaire-card-interface.ts
@@ -1,11 +1,0 @@
-export interface Card {
-  id: string;
-  type: "clubs" | "diamonds" | "hearts" | "spades";
-  black: boolean;
-  number: number;
-  facingUp: boolean;
-  place: "waste" | "foundation" | "tableau";
-  column: number;
-  backgroundPositionFacingUp?: string;
-  backgroundPositionFacingDown?: string;
-}
