--- lib/solitaire-global.ts
@@ -1,3 +1,0 @@
-import type { Card } from "~/lib/solitaire-card-interface";
-
-export const stock: Card[] = [];
